+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## 1.2.26
-- Add animated cape support
-- Fix player preview disappearing on pop-ups
+## 1.2.29
+- New panorama for 26.3
+- Disable "Block Useless Perks" in IQ Addons by default for new and existing users (if you turn it back on it will stay enabled)